Extract default homepage into plugin
